% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algos.R
\name{my.ets}
\alias{my.ets}
\title{Make a prediction with ETS algorithm for one year after last oberved point
TBATS differ from BATS in the way it models the seasonality}
\usage{
my.ets(prepedTS)
}
\arguments{
\item{prepedTS}{A list created by the \code{prepare.ts()} function}
}
\value{
A dataframe with 4 columns : date, average prediction, upper and lower 95% confidence interval bounds
}
\description{
Make a prediction with ETS algorithm for one year after last oberved point
TBATS differ from BATS in the way it models the seasonality
}
