% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_metrics.R
\name{my.mae}
\alias{my.mae}
\title{Custom (internal) function for MAE}
\usage{
my.mae(true, predicted)
}
\arguments{
\item{true}{num vector of actual values}

\item{predicted}{num vector of predicted values}
}
\value{
Num value with MAE
}
\description{
Custom (internal) function for MAE
}
