% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_metrics.R
\name{my.rmse}
\alias{my.rmse}
\title{Custom (internal) function for RMSE (not found in basic packages)}
\usage{
my.rmse(true, predicted)
}
\arguments{
\item{true}{num vector of actual values}

\item{predicted}{num vector of predicted values}
}
\value{
Num value with RMSE
}
\description{
Custom (internal) function for RMSE (not found in basic packages)
}
