% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algos.R
\name{my.tbats}
\alias{my.tbats}
\title{Make a prediction with TBATS algorithm for one year after last oberved point}
\usage{
my.tbats(prepedTS)
}
\arguments{
\item{prepedTS}{A list created by the \code{prepare.ts()} function}
}
\value{
A dataframe with 4 columns : date, average prediction, upper and lower 95% confidence interval bounds
}
\description{
Make a prediction with TBATS algorithm for one year after last oberved point
}
