% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interface.R
\name{runUserInterface}
\alias{runUserInterface}
\title{Lauches the shiny demo app for testing the package from a simple CSV file}
\usage{
runUserInterface()
}
\description{
Lauches the shiny demo app for testing the package from a simple CSV file
}
