% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.R
\name{getBestModel}
\alias{getBestModel}
\title{Implement selected algorithms of the package, train them without the last observed year
(or n_test number of points), make a prediction, and returns the best performing algorithm on this period.
If you want to save time, skip the SARIMA method.}
\usage{
getBestModel(dates, values, freq, complete = 0, n_test = NA,
  graph = T, algos = list(prophet = my.prophet, ets = my.ets, sarima =
  my.sarima, tbats = my.tbats, bats = my.bats, stlm = my.stlm, shortterm =
  my.shortterm))
}
\arguments{
\item{dates}{A vector of dates that can be parsed by lubridate}

\item{values}{A vector of same size as \code{dates}}

\item{freq}{A chacracter string that indicates the frequency of the time series ("week", "month", "quarter", "day").}

\item{complete}{A numerical value (or NA) to fill the missing data points}

\item{n_test}{number of data points to keep aside for the test (default : one year)}

\item{graph}{A boolean, if TRUE, comparison of algorithms is plotted}

\item{algos}{A list containing the algorithms to test}
}
\value{
A list contraining a character string with the name of the best method,
a gg object with the comparison between algorithms and a dataframe with predictions of all tried algorithms
}
\description{
Implement selected algorithms of the package, train them without the last observed year
(or n_test number of points), make a prediction, and returns the best performing algorithm on this period.
If you want to save time, skip the SARIMA method.
}
