% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.R
\name{my.predictions}
\alias{my.predictions}
\title{Make predictions with desired algorithms, combine them with observed data and gather everything
in one final dataframe. If you want to save time, skip the sarima algorithm which is much slower}
\usage{
my.predictions(prepedTS, algos = list(prophet = my.prophet, ets = my.ets,
  sarima = my.sarima))
}
\arguments{
\item{prepedTS}{A list created by the \code{prepare.ts()} function}

\item{algos}{A list containing the algorithms to be implemented}
}
\value{
A dataframe containing : date, actual observed values, one column per used algorithm, and
a column indicating the type of measure (mean prediction, upper or lower bound of CI)
}
\description{
Make predictions with desired algorithms, combine them with observed data and gather everything
in one final dataframe. If you want to save time, skip the sarima algorithm which is much slower
}
