% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algos.R
\name{my.shortterm}
\alias{my.shortterm}
\title{Make a prediction only upon previous year data. The prediction is built with last known smoothed evolution, previous year's
seasonal components and the last yearly cumulated value. No confidence interval is provided with this method}
\usage{
my.shortterm(prepedTS, smooth_window = 2)
}
\arguments{
\item{prepedTS}{A list created by the \code{prepare.ts()} function}
}
\value{
A dataframe with 4 columns : date, average prediction, upper and lower 95% confidence interval bounds
}
\description{
Make a prediction only upon previous year data. The prediction is built with last known smoothed evolution, previous year's
seasonal components and the last yearly cumulated value. No confidence interval is provided with this method
}
