% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_ts.R
\name{prepare.ts}
\alias{prepare.ts}
\title{Creates a list with the time series in a dataframe and a ts object, and the frequency stored
in decimal and litteral values. The result is meant to be put in the prophet or forecast functions}
\usage{
prepare.ts(dates, values, freq, complete = 0)
}
\arguments{
\item{dates}{A vector of dates that can be parsed by lubridate}

\item{values}{A vector of same size as \code{dates}}

\item{freq}{A chacracter string that indicates the frequency of the time series ("week", "month", "quarter", "day").}

\item{complete}{A numerical value (or NA) to fill the missing data points}
}
\value{
A list containing : a dataframe, a ts vector for the time series, and 2 scalars for its frequency
}
\description{
Creates a list with the time series in a dataframe and a ts object, and the frequency stored
in decimal and litteral values. The result is meant to be put in the prophet or forecast functions
}
