% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interface.R
\name{runUserInterface}
\alias{runUserInterface}
\title{Demo graphical user interface}
\usage{
runUserInterface()
}
\description{
A shiny application that allows the user to load a properly formated CSV file,
benchmark the algorithms, make a prediction and download the results. Requires additional packages
shiny, shinycssloaders, tidyr and plotly to be installed
}
\examples{
autoTS::runUserInterface()

}
